Mainly worked on logging in and registering, some work on the views too

+++ MP40.MVC/Controllers/BaseController.cs
@@ -7,7 +7,7 @@ using MP40.MVC.Models;
 
 namespace MP40.MVC.Controllers
 {
-	public class BaseController<TModel> : Controller where TModel : class, IViewModel
+	public abstract class BaseController<TModel> : Controller where TModel : class, IViewModel
 	{
 		protected IBijectiveMapper<MvcMapperProfile> mapper;
 		protected IDataService dataService;
@@ -18,71 +18,100 @@ namespace MP40.MVC.Controllers
 			this.dataService = dataService;
 		}
 
+		protected abstract bool IsUserAuthorized();
+
 		private Page<TModel> GetPage(Page page)
 		{
 			Type mappedType = mapper.GetMappedType(typeof(TModel));
 			return dataService.GetPageAndMap(mappedType, page, model => mapper.Map<TModel>(model))!;
 		}
 
-		public IActionResult Index()
+		public virtual IActionResult Index()
 		{
+			if (!IsUserAuthorized())
+				return RedirectToAction("Login", "Login");
+
 			Page<TModel> page = GetPage(new Page(1));
 
 			return View(page);
 		}
 
-		public IActionResult IndexPartial(int pageIndex)
+		public virtual IActionResult IndexPartial(int pageIndex)
 		{
-			Page<TModel> page = GetPage(new Page(pageIndex));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Page<TModel> page = GetPage(new Page(pageIndex));
 
 			return PartialView(page);
 		}
 
-		public IActionResult Details(int id)
+        public virtual IActionResult Details(int id)
 		{
-			Type mappedType = mapper.GetMappedType(typeof(TModel));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Type mappedType = mapper.GetMappedType(typeof(TModel));
 			return View(mapper.Map<TModel>(dataService.GetById(mappedType, id)));
 		}
 
-		public IActionResult Create()
+		public virtual IActionResult Create()
 		{
-			return View();
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            return View();
 		}
 
 		[HttpPost]
-		public IActionResult Create(TModel model)
+		public virtual IActionResult Create(TModel model)
 		{
-			Type mappedType = mapper.GetMappedType(typeof(TModel));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Type mappedType = mapper.GetMappedType(typeof(TModel));
 			var mappedModel = mapper.Map(mappedType, model);
 			dataService.Create(mappedType, mappedModel!);
 			return RedirectToAction(nameof(Index));
 		}
 
-		public IActionResult Edit(int id)
+		public virtual IActionResult Edit(int id)
 		{
-			Type mappedType = mapper.GetMappedType(typeof(TModel));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Type mappedType = mapper.GetMappedType(typeof(TModel));
 			return View(mapper.Map<TModel>(dataService.GetById(mappedType, id)));
 		}
 
 		[HttpPost]
-		public IActionResult Edit(int id, TModel model)
+		public virtual IActionResult Edit(int id, TModel model)
 		{
-			Type mappedType = mapper.GetMappedType(typeof(TModel));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Type mappedType = mapper.GetMappedType(typeof(TModel));
 			object? mappedModel = mapper.Map(mappedType, model);
 			dataService.Edit(mappedType, id, mappedModel!);
 			return RedirectToAction(nameof(Index));
 		}
 
-		public IActionResult Delete(int id)
+		public virtual IActionResult Delete(int id)
 		{
-			Type mappedType = mapper.GetMappedType(typeof(TModel));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Type mappedType = mapper.GetMappedType(typeof(TModel));
 			return View(mapper.Map<TModel>(dataService.GetById(mappedType, id)));
 		}
 
 		[HttpPost]
-		public IActionResult Delete(int id, TModel model)
+		public virtual IActionResult Delete(int id, TModel model)
 		{
-			Type mappedType = mapper.GetMappedType(typeof(TModel));
+            if (!IsUserAuthorized())
+                return RedirectToAction("Login", "Login");
+
+            Type mappedType = mapper.GetMappedType(typeof(TModel));
 			dataService.Delete(mappedType, id);
 			return RedirectToAction(nameof(Index));
 		}
